+++ test/phylogenies.js
@@ -248,6 +248,21 @@ describe('PhylogenyWrapper', function () {
   });
 
   describe('#asJSONLD', function () {
+    it('should preserve an existing @id on input phylogenies', function () {
+      const jsonld = new phyx.PhylogenyWrapper({
+        '@id': '#providedId',
+        newick: '((Homo_sapiens, Panthera_tigris), Mus_musculus)',
+      }).asJSONLD('#phylogeny0');
+      expect(jsonld).to.have.property('@id');
+      expect(jsonld['@id']).to.equal('#providedId');
+    });
+    it('should generate a new @id on input phylorefs', function () {
+      const jsonld = new phyx.PhylogenyWrapper({
+        newick: '((Homo_sapiens, Panthera_tigris), Mus_musculus)',
+      }).asJSONLD('#phylogeny0');
+      expect(jsonld).to.have.property('@id');
+      expect(jsonld['@id']).to.equal('#phylogeny0');
+    });
     it('should generate the phylogeny in JSON-LD as expected', function () {
       const expectedResults = [
         {
@@ -380,8 +395,11 @@ describe('PhylogenyWrapper', function () {
       ];
 
       expectedResults.forEach((expected) => {
-        const wrapper = new phyx.PhylogenyWrapper({ newick: expected.newick });
-        expect(wrapper.asJSONLD('#')).to.deep.equal(expected.jsonld);
+        const wrapper = new phyx.PhylogenyWrapper({
+          '@id': '#',
+          newick: expected.newick,
+        });
+        expect(wrapper.asJSONLD('#phylogeny0')).to.deep.equal(expected.jsonld);
       });
     });
   });